fix: move graph helpers to src/lib/graph to prevent oclif auto-discovery
oclif was treating graph/parser.ts, graph/analyzer.ts etc as subcommands.
Moved helper modules to src/lib/graph/, keeping only the command in
src/commands/graph/index.ts.
