+++ src/commands/graph.test.ts
@@ -2,8 +2,8 @@
  * Tests for graph command --days filtering and bar chart data
  */
 import { describe, it, expect } from "vitest";
-import { calculateCutoffMs, filterByDays } from "./graph/parser.js";
-import { analyzeSession } from "./graph/analyzer.js";
+import { calculateCutoffMs, filterByDays } from "../lib/graph/parser.js";
+import { analyzeSession } from "../lib/graph/analyzer.js";
 
 describe("graph --days filtering", () => {
   describe("calculateCutoffMs", () => {
