+++ src/lib/graph/analyzer.test.ts
