+++ src/lib/graph/analyzer.ts
